pyproject.toml
[project]
name = "code-quality-agent"
version = "0.1.0"
description = "AI-powered code quality intelligence agent"
requires-python = ">=3.9"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "google-generativeai>=0.3.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "radon>=6.0.1",
    "bandit>=1.7.5",
    "esprima>=4.0.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "gitpython>=3.1.40",
    "python-dotenv>=1.0.0",
    "openai>=1.35.3",
    "notion-client>=2.5.0",
    "aiohttp>=3.12.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "cli"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]
